Reduce storage requirements for spoiler log by re-using repeated string data, allowing higher sphere / item limits

+++ code/src/spoiler_data.c
@@ -2,12 +2,12 @@
 
 SpoilerData gSpoilerData = {0};
 
-char *SpoilerData_GetItemLocationString(u8 itemIndex)
+char *SpoilerData_GetItemLocationString(u16 itemIndex)
 {
     return &gSpoilerData.StringData[gSpoilerData.ItemLocations[itemIndex].LocationOffset];
 }
 
-char *SpoilerData_GetItemNameString(u8 itemIndex)
+char *SpoilerData_GetItemNameString(u16 itemIndex)
 {
     return &gSpoilerData.StringData[gSpoilerData.ItemLocations[itemIndex].ItemOffset];
 }
